chore: 기존 컴포넌트 test 페이지로 이동
